refac: Use standalone conda wrapper binary instead of symlink to ana
Instead of having ana detect when it's invoked as "conda" via symlink,
we now compile a standalone wrapper binary (src/wrappers/conda.rs) that
gets embedded into ana and written to ~/.ana/bin/conda on install.

Benefits:
- Cleaner separation of concerns
- Same approach on all platforms
- Wrapper is self-contained and independently testable
- No need for ana to detect invocation context

The wrapper binary is ~500KB and includes all the conda-spawn UX:
- Intercepts activate/deactivate/init with helpful messages
- Aliases shell -> spawn
- Filters create output to show spawn instructions
- Shows feedback hint on errors
