--- src/tools/conda_wrapper.rs
@@ -1,429 +1,0 @@
-//! Conda wrapper that provides conda-spawn based activation.
-//!
-//! This module implements a wrapper around conda that:
-//! - Intercepts `activate`, `deactivate`, and `init` commands with helpful messaging
-//! - Filters `create` output to show conda-spawn instructions
-//! - Passes through all other commands to the real conda binary
-
-use std::io::{BufRead, BufReader, Write};
-#[cfg(unix)]
-use std::path::Path;
-use std::process::{Command, Stdio};
-
-use console::style;
-
-use crate::paths;
-use crate::tools::specs::binaries;
-use crate::ui::status;
-
-/// Environment variable set by the Windows shim to indicate wrapper invocation.
-/// The shim sets this to the tool name (e.g., "conda") when invoking ana.exe as a wrapper.
-#[cfg(windows)]
-const WRAPPER_INVOCATION_ENV_VAR: &str = "_ANA_INTERNAL_WRAPPER_INVOCATION";
-
-/// Run the conda wrapper.
-///
-/// This is the entry point when ana is invoked as `conda` (via symlink or binary name).
-pub fn run(args: &[String]) -> i32 {
-    let first_arg = args.first().map(|s| s.as_str());
-
-    // Handle disabled commands
-    match first_arg {
-        Some("activate") | Some("deactivate") => {
-            print_disabled_shell_command(first_arg.unwrap());
-            return 1;
-        }
-        Some("init") => {
-            print_disabled_init();
-            return 1;
-        }
-        _ => {}
-    }
-
-    // Handle "shell" as an alias for "spawn" (like conda-express)
-    if first_arg == Some("shell") {
-        let mut new_args = args.to_vec();
-        new_args[0] = "spawn".to_string();
-        return hand_off_to_conda(&new_args);
-    }
-
-    // Check if this is a create command that needs output filtering
-    if should_filter_create_output(args) {
-        return run_conda_filtered(args);
-    }
-
-    // Pass through to conda
-    hand_off_to_conda(args)
-}
-
-/// Print message for disabled shell commands (activate/deactivate).
-fn print_disabled_shell_command(command: &str) {
-    eprintln!(
-        "{} `conda {command}` is not available via ana.",
-        style("!").yellow().bold()
-    );
-    eprintln!();
-    eprintln!("  ana uses conda-spawn for environment activation.");
-    eprintln!("  Instead of `conda activate myenv`, run:");
-    eprintln!();
-    eprintln!("    {}", style("conda shell myenv").green());
-    eprintln!();
-    eprintln!("  To leave the environment, exit the subshell (Ctrl+D or `exit`).");
-    eprintln!();
-    eprintln!("  Learn more: https://github.com/conda-incubator/conda-spawn");
-}
-
-/// Print message for disabled init command.
-fn print_disabled_init() {
-    eprintln!(
-        "{} `conda init` is not needed with ana.",
-        style("!").yellow().bold()
-    );
-    eprintln!();
-    eprintln!("  ana provides conda without requiring shell initialization.");
-    eprintln!(
-        "  Just add {} to your PATH and you're ready to go.",
-        style("~/.ana/bin").cyan()
-    );
-    eprintln!();
-    eprintln!("  To activate environments, use:");
-    eprintln!();
-    eprintln!("    {}", style("conda shell myenv").green());
-    eprintln!();
-    eprintln!("  Learn more: https://github.com/conda-incubator/conda-spawn");
-}
-
-/// Print a hint about reporting issues when conda encounters an error.
-fn print_error_feedback_hint() {
-    eprintln!();
-    eprintln!(
-        "If this error is related to ana's conda integration, please report it with {}.",
-        status::highlight("ana self feedback")
-    );
-}
-
-/// Check if this is a create command that should have filtered output.
-fn should_filter_create_output(args: &[String]) -> bool {
-    // Check for `conda create` or `conda env create`
-    let is_create = args.first().map(|s| s == "create").unwrap_or(false);
-    let is_env_create = args.len() >= 2 && args[0] == "env" && args[1] == "create";
-
-    if !is_create && !is_env_create {
-        return false;
-    }
-
-    // Only filter if -y/--yes is present or stdin is not a terminal
-    // (This avoids issues with interactive prompts)
-    let has_yes_flag = args.iter().any(|a| a == "-y" || a == "--yes");
-    let stdin_not_tty = !atty::is(atty::Stream::Stdin);
-
-    has_yes_flag || stdin_not_tty
-}
-
-/// Run conda and filter the output for create commands.
-fn run_conda_filtered(args: &[String]) -> i32 {
-    let conda_bin = get_conda_bin();
-    let prefix = paths::tool_prefix("conda");
-
-    let mut cmd = Command::new(&conda_bin);
-    cmd.args(args);
-    cmd.stdout(Stdio::piped());
-    cmd.stderr(Stdio::inherit());
-
-    // Set CONDA_ROOT_PREFIX so conda knows where its root environment is
-    cmd.env("CONDA_ROOT_PREFIX", &prefix);
-
-    let mut child = match cmd.spawn() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("Failed to run conda: {}", e);
-            return 1;
-        }
-    };
-
-    // Filter stdout
-    if let Some(stdout) = child.stdout.take() {
-        let reader = BufReader::new(stdout);
-        let mut stdout_handle = std::io::stdout().lock();
-
-        let mut in_activation_hint = false;
-
-        // Try to extract environment name from args
-        let env_name = extract_env_name(args);
-
-        for line in reader.lines() {
-            let line = match line {
-                Ok(l) => l,
-                Err(_) => continue,
-            };
-
-            // Detect conda's activation hint section
-            if line.contains("To activate this environment") {
-                in_activation_hint = true;
-                continue;
-            }
-
-            // Skip lines that are part of conda's activation instructions
-            if in_activation_hint {
-                if line.contains("conda activate") || line.contains("conda deactivate") {
-                    continue;
-                }
-                // End of activation hint block
-                if line.trim().is_empty() || !line.starts_with('#') {
-                    in_activation_hint = false;
-                    // Print our replacement message
-                    print_activation_hint(&env_name);
-                }
-            }
-
-            if !in_activation_hint {
-                let _ = writeln!(stdout_handle, "{}", line);
-            }
-        }
-    }
-
-    match child.wait() {
-        Ok(status) => status.code().unwrap_or(1),
-        Err(e) => {
-            eprintln!("Failed to wait for conda: {}", e);
-            1
-        }
-    }
-}
-
-/// Extract environment name from create args.
-fn extract_env_name(args: &[String]) -> Option<String> {
-    let mut iter = args.iter();
-    while let Some(arg) = iter.next() {
-        if arg == "-n" || arg == "--name" {
-            return iter.next().cloned();
-        }
-        if let Some(name) = arg.strip_prefix("-n=") {
-            return Some(name.to_string());
-        }
-        if let Some(name) = arg.strip_prefix("--name=") {
-            return Some(name.to_string());
-        }
-    }
-    None
-}
-
-/// Print our replacement activation hint.
-fn print_activation_hint(env_name: &Option<String>) {
-    let name = env_name.as_deref().unwrap_or("<env-name>");
-    println!("#");
-    println!("# To activate this environment, use");
-    println!("#     $ conda shell {}", name);
-    println!("#");
-    println!("# To leave the environment, exit the subshell (Ctrl+D or `exit`).");
-    println!("#");
-}
-
-/// Get the path to the real conda binary.
-fn get_conda_bin() -> std::path::PathBuf {
-    let mut conda_bin = binaries("conda")
-        .and_then(|b: Vec<std::path::PathBuf>| b.into_iter().next())
-        .expect("conda tool should have binaries defined");
-    if cfg!(windows) {
-        conda_bin.set_extension("exe");
-    }
-    paths::tool_prefix("conda").join(conda_bin)
-}
-
-/// Hand off to conda, spawning the process and waiting for completion.
-///
-/// We use spawn+wait rather than exec so we can intercept the exit code
-/// and show a reminder about reporting issues if conda fails.
-fn hand_off_to_conda(args: &[String]) -> i32 {
-    let conda_bin = get_conda_bin();
-
-    if !conda_bin.exists() {
-        eprintln!("Conda is not installed. Run `ana tool install conda` first.");
-        return 1;
-    }
-
-    let prefix = paths::tool_prefix("conda");
-
-    let mut cmd = Command::new(&conda_bin);
-    #[cfg(windows)]
-    cmd.env_remove(WRAPPER_INVOCATION_ENV_VAR);
-    cmd.args(args);
-
-    // Set CONDA_ROOT_PREFIX so conda knows where its root environment is
-    cmd.env("CONDA_ROOT_PREFIX", &prefix);
-
-    // Add ana's bin directory to PATH so our wrapper takes precedence for subcommands
-    let bin_dir = paths::bin_dir();
-    let path = std::env::var("PATH").unwrap_or_default();
-    let new_path = format!("{}:{}", bin_dir.display(), path);
-    cmd.env("PATH", new_path);
-
-    match cmd.status() {
-        Ok(status) => {
-            let code = status.code().unwrap_or(1);
-            if code != 0 {
-                print_error_feedback_hint();
-            }
-            code
-        }
-        Err(e) => {
-            eprintln!("Failed to run conda: {}", e);
-            print_error_feedback_hint();
-            1
-        }
-    }
-}
-
-/// Check if the current binary is being invoked as "conda".
-pub fn is_conda_invocation() -> bool {
-    #[cfg(unix)]
-    {
-        std::env::args()
-            .next()
-            .and_then(|arg0| Path::new(&arg0).file_name().map(|name| name == "conda"))
-            .unwrap_or(false)
-    }
-    #[cfg(windows)]
-    {
-        std::env::var(WRAPPER_INVOCATION_ENV_VAR)
-            .map(|val| val == "conda")
-            .unwrap_or(false)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_should_filter_create_output_with_yes_flag() {
-        let args: Vec<String> = vec!["create", "-n", "myenv", "-y"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert!(should_filter_create_output(&args));
-    }
-
-    #[test]
-    fn test_should_filter_create_output_with_yes_long_flag() {
-        let args: Vec<String> = vec!["create", "-n", "myenv", "--yes"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert!(should_filter_create_output(&args));
-    }
-
-    #[test]
-    fn test_should_not_filter_create_without_yes() {
-        // When running in tests, stdin is not a terminal, so this would still filter
-        // But we're testing the logic path here
-        let args: Vec<String> = vec!["create", "-n", "myenv"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        // In test environment, stdin is typically not a TTY, so this returns true
-        // The key logic being tested is that create is detected
-        let _ = should_filter_create_output(&args);
-    }
-
-    #[test]
-    fn test_should_filter_env_create() {
-        let args: Vec<String> = vec!["env", "create", "-f", "environment.yml", "-y"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert!(should_filter_create_output(&args));
-    }
-
-    #[test]
-    fn test_should_not_filter_other_commands() {
-        let args: Vec<String> = vec!["install", "numpy", "-y"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert!(!should_filter_create_output(&args));
-
-        let args: Vec<String> = vec!["list"].into_iter().map(String::from).collect();
-        assert!(!should_filter_create_output(&args));
-
-        let args: Vec<String> = vec!["info"].into_iter().map(String::from).collect();
-        assert!(!should_filter_create_output(&args));
-    }
-
-    #[test]
-    fn test_extract_env_name_short_flag() {
-        let args: Vec<String> = vec!["create", "-n", "myenv", "python"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert_eq!(extract_env_name(&args), Some("myenv".to_string()));
-    }
-
-    #[test]
-    fn test_extract_env_name_long_flag() {
-        let args: Vec<String> = vec!["create", "--name", "myenv", "python"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert_eq!(extract_env_name(&args), Some("myenv".to_string()));
-    }
-
-    #[test]
-    fn test_extract_env_name_short_equals() {
-        let args: Vec<String> = vec!["create", "-n=myenv", "python"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert_eq!(extract_env_name(&args), Some("myenv".to_string()));
-    }
-
-    #[test]
-    fn test_extract_env_name_long_equals() {
-        let args: Vec<String> = vec!["create", "--name=myenv", "python"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert_eq!(extract_env_name(&args), Some("myenv".to_string()));
-    }
-
-    #[test]
-    fn test_extract_env_name_not_present() {
-        let args: Vec<String> = vec!["create", "-p", "/path/to/env", "python"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        assert_eq!(extract_env_name(&args), None);
-    }
-
-    #[test]
-    fn test_extract_env_name_empty_args() {
-        let args: Vec<String> = vec![];
-        assert_eq!(extract_env_name(&args), None);
-    }
-
-    #[cfg(windows)]
-    mod windows_tests {
-        use super::*;
-
-        #[test]
-        fn test_is_conda_invocation_with_env_var() {
-            temp_env::with_var(WRAPPER_INVOCATION_ENV_VAR, Some("conda"), || {
-                assert!(is_conda_invocation());
-            });
-        }
-
-        #[test]
-        fn test_is_conda_invocation_with_other_tool() {
-            temp_env::with_var(WRAPPER_INVOCATION_ENV_VAR, Some("mamba"), || {
-                assert!(!is_conda_invocation());
-            });
-        }
-
-        #[test]
-        fn test_is_conda_invocation_without_env_var() {
-            temp_env::with_var(WRAPPER_INVOCATION_ENV_VAR, None::<&str>, || {
-                assert!(!is_conda_invocation());
-            });
-        }
-    }
-}
